DEVHUB-611: Fix styles and move function
